Layout: Use 5 columns, merge ack/downtime into handled, problems list is 2 rows
This better uses the space on large dashboards with more columns.

Furthermore we can just drop the ack/downtime simplemon widgets
as the problem counts really deal with handled problems only.

Making the problems list a 2 row widget is also reasonable.
I've seen that with a task list at the NETWAYS dashboards too.