Move Logo
The icinga logo masthead looked a bit inconsistent when floating around separated
from the actual grid.  I've moved it so it is relative to the top left hand corner
of the grid.  Requires #38 so the style doesn't get broken (e.g. requires the
background-color of the body not to be set to white.