Allow to configure the Icinga Web 2 Iframe URL
This is a little brainfuck, as we need to create
an instance of the icinga2 class for reading the configuration
details in `config.ru`.

We'll expose the setting inside a helper function, which is
called inside the HTML erb template for the dashboard.

Note: This is done every time the user presses refresh. Job execution
won't trigger this. The previous attempt with updating this on send_event
caused flickering.